+++ spec/unit/inspector_spec.rb
@@ -19,8 +19,14 @@ module Capybara::Poltergeist
 
     it 'can be opened' do
       subject = Inspector.new('chromium', 1234)
-      Process.should_receive(:spawn).with('chromium', '//localhost:1234/')
+      Process.should_receive(:spawn).with('chromium', 'http://localhost:1234/')
       subject.open
+    end
+
+    it 'can be opened with https URL' do
+      subject = Inspector.new('chromium', 1234)
+      Process.should_receive(:spawn).with('chromium', 'https://localhost:1234/')
+      subject.open 'https'
     end
 
     it 'raises an error on open when the browser is unknown' do